backend/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
pydantic==2.7.1
python-dotenv==1.0.1
openai==1.30.1
httpx==0.27.0
pytest==8.2.2
pytest-asyncio==0.23.7
anyio==4.4.0
redis==6.2.0
slowapi==0.1.9
